BlockSnap/requirements.txt
# Core dependencies
Flask==3.0.0
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.1.0
python-json-logger==2.0.7

# Blockchain dependencies
web3==6.11.1
eth-account==0.9.0
eth-utils==2.3.0

# Image processing (for mock camera)
opencv-python==4.8.1.78
numpy==1.24.3

# Optional Raspberry Pi dependencies (comment out when testing on other platforms)
# RPi.GPIO==0.7.1
# picamera2==0.3.12
# gps==3.19